requirements.txt
torch==1.10.1
torchvision==0.11.2
numpy==1.21.4
matplotlib==3.4.3
Pillow==8.4.0
IPython==7.30.1